- use the global "startup_time" var instead of local ones
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@5839 689a6050-402a-0410-94f2-e92a70836424

+++ modules/presence/presentity.c
@@ -64,7 +64,8 @@ char* generate_ETag(int publ_count)
 		ERR_MEM(PKG_MEM_STR);
 	}
 	memset(etag, 0, ETAG_LEN*sizeof(char));
-	size = sprintf (etag, "%c.%d.%d.%d.%d",prefix, startup_time, pid, counter, publ_count);
+	size = sprintf (etag, "%c.%d.%d.%d.%d",
+		prefix, (int)startup_time, pid, counter, publ_count);
 	if( size <0 )
 	{
 		LM_ERR("unsuccessfull sprintf\n ");
